Moved admin book pages to separate folders
