+++ src/pages/admin-panel/pages/book/pages/editBook/EditBook.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const EditBook = () => {
+  return <div>Edit Book</div>;
+};
+
+export default EditBook;
